Spring implemented. Console divided. Test written for new methods implemented.
