+++ pom.xml
@@ -28,7 +28,7 @@
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-context</artifactId>
-            <version>4.2.2.RELEASE</version>
+            <version>5.2.7.RELEASE</version>
         </dependency>
         <dependency>
             <groupId>org.jetbrains</groupId>
