+++ src/main/java/com/javaguru/shoppinglist/ShoppingListApplication.java
@@ -1,0 +1,15 @@
+package com.javaguru.shoppinglist;
+
+import com.javaguru.shoppinglist.config.AppConfig;
+import com.javaguru.shoppinglist.console.Console;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+
+class ShoppingListApplication {
+
+    public static void main(String[] args) {
+        ApplicationContext context = new AnnotationConfigApplicationContext(AppConfig.class);
+        Console console = context.getBean(Console.class);
+        console.execute();
+    }
+}
