+++ src/main/java/com/javaguru/shoppinglist/config/AppConfig.java
@@ -1,0 +1,11 @@
+package com.javaguru.shoppinglist.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan(basePackages = "com.javaguru.shoppinglist")
+public class AppConfig {
+
+
+}
