+++ src/main/java/com/javaguru/shoppinglist/console/ExitAction.java
@@ -1,0 +1,20 @@
+package com.javaguru.shoppinglist.console;
+
+import org.springframework.core.annotation.Order;
+import org.springframework.stereotype.Component;
+
+@Component
+@Order(6)
+public class ExitAction implements MenuAction {
+
+    @Override
+    public void execute() {
+        System.exit(0);
+    }
+
+    @Override
+    public String toString() {
+        return "Exit";
+    }
+
+}
