+++ src/main/java/com/javaguru/shoppinglist/console/FindByIdAction.java
@@ -1,0 +1,31 @@
+package com.javaguru.shoppinglist.console;
+
+import com.javaguru.shoppinglist.service.ProductService;
+import org.springframework.core.annotation.Order;
+import org.springframework.stereotype.Component;
+
+import java.util.Scanner;
+
+@Component
+@Order(2)
+public class FindByIdAction implements MenuAction {
+
+    private final ProductService productService;
+
+    public FindByIdAction(ProductService productService) {
+        this.productService = productService;
+    }
+
+    @Override
+    public void execute() {
+        Scanner scanner = new Scanner(System.in);
+        System.out.println("Enter product id: ");
+        long id = scanner.nextLong();
+        System.out.println(productService.findByID(id));
+    }
+
+    @Override
+    public String toString() {
+        return "Find product by ID";
+    }
+}
