+++ src/main/java/com/javaguru/shoppinglist/console/MenuAction.java
@@ -1,0 +1,7 @@
+package com.javaguru.shoppinglist.console;
+
+public interface MenuAction {
+
+    void execute();
+
+}
