+++ src/main/java/com/javaguru/shoppinglist/domain/ProductEntity.java
