+++ src/main/java/com/javaguru/shoppinglist/dto/ProductDto.java
