+++ src/main/java/com/javaguru/shoppinglist/service/validation/exceptions/NameAlreadyExistsException.java
