+++ src/main/java/com/javaguru/shoppinglist/service/validation/exceptions/NameIllegalException.java
