+++ src/main/java/com/javaguru/shoppinglist/service/validation/exceptions/WrongMenuInputException.java
@@ -1,0 +1,8 @@
+package com.javaguru.shoppinglist.service.validation.exceptions;
+
+public class WrongMenuInputException extends ValidationException {
+
+    public WrongMenuInputException(String message) {
+        super(message);
+    }
+}
