+++ src/test/java/com/javaguru/shoppinglist/mappers/ProductMapperTest.java
