+++ src/test/java/com/javaguru/shoppinglist/service/validation/rules/ProductValidationRuleTest.java
